app/requirements.txt
Flask==0.12.3
requests==2.22.0
config==0.4.2
python-decouple==3.1
sentry-sdk[flask]==0.10.2
Flask-BasicAuth==0.2.0
kafka-python==1.4.6
redis
elasticsearch
elasticsearch-dsl
